backend/requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-cors-headers==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
drf-yasg==1.17.0
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.2
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pyparsing==2.4.5
pyrsistent==0.15.5
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
uritemplate==3.0.0
urllib3==1.25.7
zipp==0.6.0
